[Feat/#17] AppDownloadButton 관련 코드 분리 등
